# openapi/openapi-gpt-plugin-bridge.yaml
openapi: 3.1.0
info:
  title: Figma Gateway GPT Plugin Bridge API
  version: 0.1.0
  description: |
    GPT Actions schema for read-only alias operations and page creation through the Figma plugin bridge.

    Notes:
    - createPage requires a running Figma plugin bridge session
    - registerPluginSession is usually handled automatically by the plugin itself
    - createPage can auto-resolve the latest active plugin session when sessionId is omitted
    - fileKey is optional for createPage; the server can use the current plugin session context automatically
    - createPage is consequential and queues a command for the plugin
    - createPage defaults to live execution when dryRun is omitted
servers:
  - url: https://figma-gateway.vazovski.art
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: '#/components/schemas/Error'
    AliasRecord:
      type: object
      properties:
        alias:
          type: string
        fileKey:
          type: string
        nodeId:
          type: string
        project:
          type: string
        tags:
          type: array
          items:
            type: string
        description:
          type: string
    AliasListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/AliasRecord'
    AliasResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AliasRecord'
    ResolveAliasRequest:
      type: object
      required:
        - alias
      properties:
        alias:
          type: string
    SessionRegisterRequest:
      type: object
      properties:
        fileKey:
          type: string
          description: Optional cloud file key when available from Figma.
        localFileKey:
          type: string
          description: Optional local editor identifier when a cloud file key is not available.
        fileName:
          type: string
          description: Optional current file name for local or draft files.
        clientName:
          type: string
    SessionRegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            sessionId:
              type: string
            sessionToken:
              type: string
            fileKey:
              type: string
            clientName:
              type: string
            createdAt:
              type: string
...
paths:
...
  /api/aliases:
    get:
      operationId: listAliases
      security:
        - bearerAuth: []
      parameters:
        - name: project
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AliasListResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/resolve-alias:
    post:
      operationId: resolveAlias
      x-openai-isConsequential: false
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveAliasRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AliasResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/plugin-bridge/sessions/active:
    get:
      operationId: listActivePluginSessions
      x-openai-isConsequential: false
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/plugin-bridge/sessions/register:
    post:
      operationId: registerPluginSession
      x-openai-isConsequential: false
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionRegisterRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionRegisterResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
security:
  - bearerAuth: []